Gui: With Qt6 QColor's redF(), greenF(), blueF() methods return a float

+++ src/Gui/PreferencePages/DlgSettingsLightSources.cpp
@@ -210,10 +210,10 @@ void DlgSettingsLightSources::lightColor()
 {
     if (view) {
         QColor color = ui->light1Color->color();
-        double red = color.redF();
-        double green = color.greenF();
-        double blue = color.blueF();
-        view->getHeadlight()->color = SbColor(float(red), float(green), float(blue));
+        float red = float(color.redF());
+        float green = float(color.greenF());
+        float blue = float(color.blueF());
+        view->getHeadlight()->color = SbColor(red, green, blue);
     }
 }
